requirements.txt
# Dependencies from screenshot (with exact versions)
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0

# Project dependencies (no version pinning = latest)
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
ibm-watsonx-ai
langchain
langchain-ibm
nltk
scikit-learn
scipy
sentence-transformers
joblib
threadpoolctl
sse-starlette
starlette
matplotlib
scikit-learn
streamlit
pinecone
python-multipart